feat(api): entidades de domínio, documentação da api e cofigurações de ambiente.
