--- ProdutosApp.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35707.178 d17.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProdutosApp.Api", "ProdutosApp.Api\ProdutosApp.Api.csproj", "{42283B43-0FDB-4FDF-B186-2DD4C0E874F5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{42283B43-0FDB-4FDF-B186-2DD4C0E874F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{42283B43-0FDB-4FDF-B186-2DD4C0E874F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{42283B43-0FDB-4FDF-B186-2DD4C0E874F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{42283B43-0FDB-4FDF-B186-2DD4C0E874F5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
