+++ src/ProdutosApp.Api/Program.cs
@@ -1,0 +1,30 @@
+using Scalar.AspNetCore;
+
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddControllers();
+builder.Services.AddOpenApi();
+
+//Swagger
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+    app.MapOpenApi();
+}
+
+//Swagger
+app.UseSwagger();
+app.UseSwaggerUI();
+
+//Scalar
+app.MapScalarApiReference(options => {
+    options.WithTheme(ScalarTheme.BluePlanet);
+});
+
+app.UseAuthorization();
+app.MapControllers();
+app.Run();
