+++ src/ProdutosApp.Api/appsettings.Development.json
@@ -1,0 +1,6 @@
+{
+  /* Mapeamento do banco de dados local */
+  "ConnectionStrings": {
+    "ProdutosAppBD": "Data Source=localhost,1434;Initial Catalog=master;User ID=sa;Password=Coti@2025;Encrypt=False"
+  }
+}
