+++ src/ProdutosApp.Api/appsettings.Production.json
@@ -1,0 +1,5 @@
+﻿{
+  "ConnectionStrings": {
+
+  }
+}
