+++ src/ProdutosApp.Domain/Entities/Categoria.cs
@@ -1,0 +1,17 @@
+﻿namespace ProdutosApp.Domain.Entities;
+
+/// <summary>
+/// Modelo de entidade de domínio para Categoria.
+/// </summary>
+public class Categoria
+{
+    #region Propriedades
+    public Guid? Id { get; set; }
+    public string? Nome { get; set; }
+    public bool? Ativo { get; set; }
+    #endregion
+
+    #region Relacionamentos
+    public ICollection<Produto>? Produtos { get; set; }
+    #endregion
+}
